requirements.txt
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
pyserial>=3.5 